chore: move setup-trusted-address command to its.js
